Move sdl2 modified platform specific files out of repo

+++ visual/win32/BitmapBitsAlloc.h
@@ -1,7 +1,6 @@
 
 #ifndef __BITMAP_BITS_ALLOC_H__
 #define __BITMAP_BITS_ALLOC_H__
-#include <stdint.h>
 
 //---------------------------------------------------------------------------
 // memory allocation class
@@ -16,7 +15,7 @@ public:
 // heap allocation functions for bitmap bits
 //---------------------------------------------------------------------------
 class tTVPBitmapBitsAlloc {
-	typedef intptr_t tTJSPointerSizedInteger;
+	typedef DWORD tTJSPointerSizedInteger;
 	// this must be a integer type that has the same size with the normal
 	// pointer ( void *)
 	//---------------------------------------------------------------------------
